Add Formula for proj version 6.3.1
Heroku-20 stack only has proj4 version 6.3.1 available and we want
to be running the same version on our development environment as we
use in production. Homebrew core only has formula for versions 7 and 8
of proj4 so we want to host our own formula for proj version 6.3.1 to
ease the process of setting up our development environment. (no
building from source)

Co-authored-by: James Nebeker <jnebeker@radiusnetworks.com>